+++ src/kedger/graph/__init__.py
@@ -1,4 +1,15 @@
 from kedger.graph.entities import upsert_entity
-from kedger.graph.expand import WalkNotebook, associative_expand, notebook_walk
+from kedger.graph.expand import (
+    WalkNotebook,
+    associative_expand,
+    notebook_walk,
+    seed_idf_scores,
+)
 
-__all__ = ["associative_expand", "notebook_walk", "WalkNotebook", "upsert_entity"]
+__all__ = [
+    "associative_expand",
+    "notebook_walk",
+    "WalkNotebook",
+    "seed_idf_scores",
+    "upsert_entity",
+]